Simple thumbnail support
Images are loaded and resized on the fly into textures via async tasks, which are then used by the grid to display thumbnails. An LRU cache is used to evict textures when the in-memory size of the thumbnails as a whole exceeds a certain limit (currently set to 512 MiB of raw image data). At the moment, the thumbnail loading logic has been nuzzled via a load delay and a mutex (boolean flag) to prevent a flood of async tasks all loading images from exploding the memory footprint of the app.